# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Kura Kaani API Documentation",
        "description": "Kura Kaani API Documentation",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/chats/{chatId}/messages": {
            "get": {
                "tags": [
                    "Message"
                ],
                "summary": "Get chat history",
                "operationId": "4f55fa24f067a681879b94244c4e44b7",
                "parameters": [
                    {
                        "name": "chatId",
                        "in": "path",
                        "description": "ID of the chat",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Chat history retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer"
                                            },
                                            "sender_id": {
                                                "type": "integer"
                                            },
                                            "receiver_id": {
                                                "type": "integer"
                                            },
                                            "message": {
                                                "type": "string"
                                            },
                                            "created_at": {
                                                "type": "string",
                                                "format": "date-time"
                                            },
                                            "updated_at": {
                                                "type": "string",
                                                "format": "date-time"
                                            },
                                            "sender": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer"
                                                    },
                                                    "name": {
                                                        "type": "string"
                                                    }
                                                },
                                                "type": "object"
                                            },
                                            "receiver": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer"
                                                    },
                                                    "name": {
                                                        "type": "string"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Not authorized to view this chat",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Not authorized to view this chat"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Message"
                ],
                "summary": "Send a message in a chat",
                "operationId": "19ef0426ac01e93d153d450878a94c7c",
                "parameters": [
                    {
                        "name": "chatId",
                        "in": "path",
                        "description": "ID of the chat",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Message data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "message"
                                ],
                                "properties": {
                                    "message": {
                                        "type": "string",
                                        "example": "Hello, how are you?"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Message sent successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "description": "ID of the sent message",
                                            "type": "integer"
                                        },
                                        "sender_id": {
                                            "description": "ID of the message sender",
                                            "type": "integer"
                                        },
                                        "receiver_id": {
                                            "description": "ID of the message receiver",
                                            "type": "integer"
                                        },
                                        "message": {
                                            "description": "Content of the message",
                                            "type": "string"
                                        },
                                        "created_at": {
                                            "description": "Timestamp when the message was sent",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Not authorized to send messages in this chat",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Not authorized to send messages in this chat"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
...
}